[API][HTML][Implementation] - Update assortment
Added code to html and API endpoint for updating an assortment's title
and description. Created a modal that pops up to edit title and
description. There is still a bug where when you edit an assortment, the
name is not automaticaly updated on the screen, and the Delete and View
Cards button do not work until you refresh.