[Security][Implementation] - Delete account
Added functionality for deleting an account from the database, and it
sends you to the /login page. A modal pops up to confirm too. Logging in
with the account will not work. I want to see if I can make it so that
instead of a button, you can press your username and have a dropdown
come up instead.

- #18