[Security] - Confirm password
Added a second field when creating a new user where you enter your
password twice to confirm that you are entering it correctly. You are
unable to register until the confirm password is correct. A message pops
up if the passwords do not match. A user is successfully made if valid.

- #17
- #21