[CSS] - Create new assortment button + profile picture
Moved the Create New Assortment to the top right corner on the same row
as the title. Also created a profile picture in the navbar which is just
a grey circle with the username's first letter capitalized. Clicking on
the profile picture sends you to /settings like when you click on your
username.

- #25