[CSS][Implementation] - Restructured assortments/id page
Restructured the assortments/id page so that instead of making a card at
the top, you create a card at the bottom as you create more cards.
Interacting with existing cards works fine. However, I encountered a bug
where when you press the Create New Card button at the bottom, a new
card will appear in the database and on localhost, but you will not be
able to type into any fields or delete the card until you refresh the
page. Also, deleting a card from an assortment on localhost will remove
it from your screen and the join table, but the card still exists in the
cards database.

- #29
- #30