[CSS][Debug] - Card working after Create New Card
Added one line (lol) to fix a bug so that when you press on the Create
New Card button, you are now able to type into the card and the fields
saved to the database. You are also now able to delete the card upon
pressing Create New Card without needing to refresh.

- #29
- #30