[Debug] - Each field height for each card being updated
Fixed a bug where the heights of each field would be updated for each
card rather than one a per-card basis. This also resolved the bug where
I could not change the heights of the complexities!!!!!!!!!!!!! This
must have been because there was a tall complexity section in a
different question.

- #29