[CSS][Implementation] - Still Learning/Know buttons
Added code so that instead of Previous/Next buttons to see the previous
and next cards, there are 2 new buttons: Still Learning to put the card
in a temporary Still Learning group, and Next to move on. When all of
the cards are done being reviewed, there will be 3 buttons that show up:
1 to review all Still Learning cards, 1 to review all cards, and 1 to
return to the assortment. If there are no cards in Still Learning, then
the button is faded out.

- #32