[CSS][Implementation] - Shuffle functionality
Added shuffle functionality to the study page. There is a button with a
shuffle icon on the study page that you can press to shuffle the cards.
WHen you get to the end of the deck and see the Still Learning deck,
those cards will still be shuffled. There is no way to unshuffle but I
do not think I will add a way.

- #32
- #33